Fix issue with setting ftw-server httpd slot in start-ftw-server!.
When the module is compiled the set! (using set! macro, or the -set!
procedure directly) fails with a error of "*** ERROR IN
ftw/server#start-ftw-server! -- Cannot find slot #<ftw-server #24>
class" after starting the httpd server.  The code works perfectly when
interpreted.

Wrapping the set! in an apply allows dynamic look-up on the slot at
runtime and prevents the issue.